Package `Icons/**/*.ico` into a buildable .dll solution
You can still access the icons exactly as you used to, however, you can
now also build this project to provide an accessible .dll that provides
the .ico files on request.

This is super useful for being able to browse through all of the
available files in an icon browser, just like you should `Shell32.dll`

In this commit, I've also removed a lot of the boilerplating within the
files and within the .md documents.

I've updated the description, readme, and credits.